Initial commit for base-carousel conversion
