+++ extensions/amp-base-carousel/1.0/component.jss.js
